+++ device-scheduler/grpalloc/grpallocate.go
@@ -3,8 +3,9 @@ package grpalloc
 import (
 	"regexp"
 
-	"github.com/Microsoft/KubeGPU/gpuextension/grpalloc/resource"
-	"github.com/Microsoft/KubeGPU/gpuextension/grpalloc/scorer"
+	"github.com/Microsoft/KubeGPU/device-scheduler/grpalloc/resource"
+	"github.com/Microsoft/KubeGPU/device-scheduler/grpalloc/scorer"
+	sctypes "github.com/Microsoft/KubeGPU/device-scheduler/types"
 	"github.com/Microsoft/KubeGPU/types"
 	"github.com/Microsoft/KubeGPU/utils"
 	"github.com/golang/glog"
@@ -137,7 +138,7 @@ func (grp *GrpAllocator) resetGroup(restorePoint *GrpAllocator) {
 // returns whether resource available and score
 // can use hash of scorers for different resources
 // simple leftover for now for score, 0 is low, 1.0 is high score
-func (grp *GrpAllocator) resourceAvailable(resourceLocation string) (bool, []types.PredicateFailureReason) {
+func (grp *GrpAllocator) resourceAvailable(resourceLocation string) (bool, []sctypes.PredicateFailureReason) {
 	grpAllocRes := grp.GrpAllocResource[resourceLocation]
 
 	glog.V(5).Infoln("Resource requirments")
@@ -146,7 +147,7 @@ func (grp *GrpAllocator) resourceAvailable(resourceLocation string) (bool, []typ
 	printResMap(grp.AllocResource, grpAllocRes, grp.IsAllocSubGrp)
 
 	found := true
-	var predicateFails []types.PredicateFailureReason
+	var predicateFails []sctypes.PredicateFailureReason
 	for grpReqKey, grpReqElem := range grp.GrpRequiredResource {
 		if !grp.IsReqSubGrp[grpReqKey] {
 			// see if resource exists
@@ -193,10 +194,10 @@ func (grp *GrpAllocator) allocateSubGroups(
 	allocLocationName string,
 	subgrpsReq map[string](map[string](map[string]string)),
 	subgrpsAllocRes map[string](map[string](map[string]string))) (
-	bool, []types.PredicateFailureReason) {
+	bool, []sctypes.PredicateFailureReason) {
 
 	found := true
-	var predicateFails []types.PredicateFailureReason
+	var predicateFails []sctypes.PredicateFailureReason
 	sortedSubGrpsReqKeys := utils.SortedStringKeys(subgrpsReq)
 	for _, subgrpsKey := range sortedSubGrpsReqKeys {
 		subgrpsElemGrp := subgrpsReq[subgrpsKey]
@@ -218,9 +219,9 @@ func (grp *GrpAllocator) allocateSubGroups(
 	return found, predicateFails
 }
 
-func (grp *GrpAllocator) findScoreAndUpdate(location string) (bool, []types.PredicateFailureReason) {
+func (grp *GrpAllocator) findScoreAndUpdate(location string) (bool, []sctypes.PredicateFailureReason) {
 	found := true
-	var predicateFails []types.PredicateFailureReason
+	var predicateFails []sctypes.PredicateFailureReason
 
 	// first compute list of requested resources to alloctable resources
 	requestedResource := make(map[string]([]int64))
@@ -262,7 +263,7 @@ func (grp *GrpAllocator) findScoreAndUpdate(location string) (bool, []types.Pred
 }
 
 func (grp *GrpAllocator) allocateGroupAt(location string,
-	subgrpsReq map[string](map[string](map[string]string))) (bool, []types.PredicateFailureReason) {
+	subgrpsReq map[string](map[string](map[string]string))) (bool, []sctypes.PredicateFailureReason) {
 
 	allocLocationName := grp.AllocBaseGroupPrefix + "/" + location
 	grpsAllocResElem := grp.GrpAllocResource[location]
@@ -310,7 +311,7 @@ func (grp *GrpAllocator) allocateGroupAt(location string,
 // usedResource is the amount of utilized resource in the global resource list
 // i.e. usedResource[grpAllocRes[i][n]] is used when considering the "i"th alloctable group of the "n"th resource
 // i.e. usedResource[allocated[grpReq[n]]] is subtracted from after allocation
-func (grp *GrpAllocator) allocateGroup() (bool, []types.PredicateFailureReason) {
+func (grp *GrpAllocator) allocateGroup() (bool, []sctypes.PredicateFailureReason) {
 	if len(grp.GrpRequiredResource) == 0 {
 		return true, nil
 	}
@@ -320,7 +321,7 @@ func (grp *GrpAllocator) allocateGroup() (bool, []types.PredicateFailureReason) 
 	maxScoreGrp := grp
 	maxIsUsedGroup := false
 	maxGroupName := ""
-	var predicateFails []types.PredicateFailureReason
+	var predicateFails []sctypes.PredicateFailureReason
 
 	// find subgroups for required resources
 	subgrpsReq, isSubGrp := findSubGroups(grp.ReqBaseGroupName, grp.GrpRequiredResource)
@@ -388,7 +389,7 @@ func containerFitsGroupConstraints(contName string, contReq *types.ContainerInfo
 	allocatable types.ResourceList, allocScorer map[string]scorer.ResourceScoreFunc,
 	podResource map[string]int64, nodeResource map[string]int64,
 	usedGroups map[string]bool, bPreferUsed bool, bSetAllocateFrom bool) (
-	*GrpAllocator, bool, []types.PredicateFailureReason, float64) {
+	*GrpAllocator, bool, []sctypes.PredicateFailureReason, float64) {
 
 	grp := &GrpAllocator{}
 
@@ -454,7 +455,7 @@ func containerFitsGroupConstraints(contName string, contReq *types.ContainerInfo
 	grp.NodeResource = nodeResource
 
 	var found bool
-	var reasons []types.PredicateFailureReason
+	var reasons []sctypes.PredicateFailureReason
 	var score float64
 
 	if contReq.AllocateFrom == nil || (len(contReq.AllocateFrom) == 0 && len(req) > 0) {
@@ -517,12 +518,12 @@ func setScoreFunc(n *types.NodeInfo) map[string]scorer.ResourceScoreFunc {
 }
 
 // PodFitsGroupConstraints tells if pod fits constraints, score returned is score of running containers
-func PodFitsGroupConstraints(n *types.NodeInfo, spec *types.PodInfo, allocating bool) (bool, []types.PredicateFailureReason, float64) {
+func PodFitsGroupConstraints(n *types.NodeInfo, spec *types.PodInfo, allocating bool) (bool, []sctypes.PredicateFailureReason, float64) {
 	podResource := make(map[string]int64)
 	nodeResource := initNodeResource(n)
 	usedGroups := make(map[string]bool)
 	totalScore := 0.0
-	var predicateFails []types.PredicateFailureReason
+	var predicateFails []sctypes.PredicateFailureReason
 	found := true
 
 	// set score function for alloctable resources at node
