+++ device-scheduler/types/typeutils.go
