+++ types/types.go
@@ -1,15 +1,13 @@
 package types
 
+type ResourceName string
+
 const (
 	// NVIDIA GPU, in devices. Alpha, might change: although fractional and allowing values >1, only one whole device per node is assigned.
 	ResourceGPU ResourceName = "alpha.gpu/numgpu"
-	// auto topology generation "0" means default (everything in its own group)
-	GPUTopologyGeneration ResourceName = "alpha.gpu/gpu-generate-topology"
 	// Namespace prefix for group resources.
 	DeviceGroupPrefix = "alpha/grpresource"
 )
-
-type ResourceName string
 
 // ResourceLocation is a set of (resource name, resource location on node) pairs.
 type ResourceLocation map[ResourceName]ResourceName
@@ -115,56 +113,6 @@ func NewNodeInfoWithName(name string) *NodeInfo {
 	return node
 }
 
-type Volume struct {
-	Name   string
-	Driver string
-}
-
-// Device is a device to use
-type Device interface {
-	// New creates the device and initializes it
-	New() error
-	// Start logically initializes the device
-	Start() error
-	// UpdateNodeInfo - updates a node info structure by writing capacity, allocatable, used, scorer
-	UpdateNodeInfo(*NodeInfo) error
-	// Allocate attempst to allocate the devices
-	// Returns list of (VolumeName, VolumeDriver), and list of Devices to use
-	// Returns an error on failure.
-	Allocate(*PodInfo, *ContainerInfo) ([]Volume, []string, error)
-	// GetName returns the name of a device
-	GetName() string
-}
-
-type PredicateFailureReason interface {
-	GetReason() string
-	GetInfo() (ResourceName, int64, int64, int64)
-}
-
-// used by scheduler
-type DeviceScheduler interface {
-	// see if pod fits on node & return device score
-	PodFitsDevice(nodeInfo *NodeInfo, podInfo *PodInfo, fillAllocateFrom bool, runGrpScheduler bool) (bool, []PredicateFailureReason, float64)
-	// allocate resources
-	PodAllocate(nodeInfo *NodeInfo, podInfo *PodInfo, runGrpScheduler bool) error
-	// take resources from node
-	TakePodResources(*NodeInfo, *PodInfo, bool) error
-	// return resources to node
-	ReturnPodResources(*NodeInfo, *PodInfo, bool) error
-	// GetName returns the name of a device
-	GetName() string
-	// Tells whether group scheduler is being used?
-	UsingGroupScheduler() bool
-}
-
-const (
-	DefaultScorer = iota // 0
-	LeftOverScorer
-	EnumLeftOverScorer
-)
-
-type SortedTreeNode struct {
-	Val   int
-	Score float64 // used for tie breaker
-	Child []*SortedTreeNode
+func AddGroupResource(list ResourceList, key string, val int64) {
+	list[ResourceName(DeviceGroupPrefix+"/"+key)] = val
 }
